cancel Pending verifications in `DataModificationApprovalObserver`

+++ plugins/txes/storage/src/config/StorageConfiguration.h
@@ -38,7 +38,7 @@ namespace catapult { namespace config {
 		bool Enabled;
 
 		/// Verification frequency.
-		uint64_t VerificationFrequency;
+		uint16_t VerificationFrequency;
 
 	private:
 		StorageConfiguration() = default;
